+++ apps/demo/contracts/CHANGELOG.md
@@ -6,4 +6,4 @@
 
 - Updated dependencies [2027d29]
   - @sundaeswap/capacity-exchange-core@1.1.0
-  - @capacity-exchange/midnight-node@1.0.1
+  - @sundaeswap/capacity-exchange-nodejs@1.0.1